refactor: isolate run platform handling
